requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3

# Firebase Auth
firebase-admin==6.3.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Cloud SQL connector (optional, for production)
cloud-sql-python-connector[asyncpg]==1.5.0
